restructured fv3run into multiple files
